remove command ^ and add :
Reason: previously when * existed, there were two rotating commands. In total 4 commands would be needed to satisfy all rotation directions, but, with the power of computers, it could be shrunken down to 3, by using a "change rotation" variable... but since we now only got 1 rotator, we could just simply add a rotator command for the other direction, which makes life simpler.
: )